+++ packages/common/src/index.ts
@@ -889,6 +889,7 @@ export type CreateProject = Omit<
     | 'createdByUserUuid'
 > & {
     warehouseConnection: CreateWarehouseCredentials;
+    copyWarehouseConnectionFromUpstreamProject?: boolean;
 };
 
 export type UpdateProject = Omit<