+++ packages/common/src/utils/formatting.ts
@@ -158,7 +158,7 @@ export const parseTimestamp = (
 function getFormatNumberOptions(value: number, format?: CustomFormat) {
     const hasCurrency =
         format?.type === CustomFormatType.CURRENCY && format?.currency;
-    const currencyOptions: Intl.NumberFormatOptions = hasCurrency
+    const currencyOptions = hasCurrency
         ? { style: 'currency', currency: format.currency }
         : {};