+++ packages/frontend/src/features/metricsCatalog/components/MetricExploreModal.tsx
@@ -41,16 +41,16 @@ import { useCatalogMetricsWithTimeDimensions } from '../hooks/useCatalogMetricsW
 import { useCatalogSegmentDimensions } from '../hooks/useCatalogSegmentDimensions';
 import { useMetric } from '../hooks/useMetricsCatalog';
 import { useRunMetricExplorerQuery } from '../hooks/useRunMetricExplorerQuery';
+import { MetricExploreComparison } from './visualization/MetricExploreComparison';
 import { MetricExploreFilter } from './visualization/MetricExploreFilter';
-import { MetricPeekComparison } from './visualization/MetricPeekComparison';
-import { MetricPeekSegmentationPicker } from './visualization/MetricPeekSegmentationPicker';
+import { MetricExploreSegmentationPicker } from './visualization/MetricExploreSegmentationPicker';
 import MetricsVisualization from './visualization/MetricsVisualization';
 
 type Props = Pick<ModalProps, 'opened' | 'onClose'> & {
     metrics: CatalogField[];
 };
 
-export const MetricPeekModal: FC<Props> = ({ opened, onClose, metrics }) => {
+export const MetricExploreModal: FC<Props> = ({ opened, onClose, metrics }) => {
     const { track } = useTracking();
 
     const organizationUuid = useAppSelector(
@@ -76,10 +76,10 @@ export const MetricPeekModal: FC<Props> = ({ opened, onClose, metrics }) => {
     const [dateRange, setDateRange] = useState<MetricExplorerDateRange | null>(
         null,
     );
-
     const [timeDimensionOverride, setTimeDimensionOverride] = useState<
         TimeDimensionConfig | undefined
     >();
+    const [filter, setFilter] = useState<FilterRule | undefined>(undefined);
 
     const resetQueryState = useCallback(() => {
         setQuery({
@@ -88,6 +88,7 @@ export const MetricPeekModal: FC<Props> = ({ opened, onClose, metrics }) => {
         });
         setTimeDimensionOverride(undefined);
         setDateRange(null);
+        setFilter(undefined);
     }, [setQuery, setTimeDimensionOverride, setDateRange]);
 
     const currentMetricIndex = useMemo(() => {
@@ -107,12 +108,12 @@ export const MetricPeekModal: FC<Props> = ({ opened, onClose, metrics }) => {
 
     const navigateToMetric = useCallback(
         (metric: CatalogField) => {
+            resetQueryState();
+
             void navigate({
                 pathname: `/projects/${projectUuid}/metrics/peek/${metric.tableName}/${metric.name}`,
                 search: location.search,
             });
-
-            resetQueryState();
         },
         [navigate, projectUuid, resetQueryState, location.search],
     );
@@ -158,8 +159,6 @@ export const MetricPeekModal: FC<Props> = ({ opened, onClose, metrics }) => {
             query.metric.table === ''
         );
     }, [query]);
-
-    const [filter, setFilter] = useState<FilterRule | undefined>(undefined);
 
     const isQueryEnabled =
         (!!projectUuid &&
@@ -522,11 +521,11 @@ export const MetricPeekModal: FC<Props> = ({ opened, onClose, metrics }) => {
                             py="md"
                         >
                             <MetricExploreFilter
-                                // TODO: Get filters from the query instead of segmentByData
                                 dimensions={availableFilters}
                                 onFilterApply={handleFilterApply}
+                                key={`${tableName}-${metricName}`}
                             />
-                            <MetricPeekSegmentationPicker
+                            <MetricExploreSegmentationPicker
                                 query={query}
                                 onSegmentDimensionChange={
                                     handleSegmentDimensionChange
@@ -586,7 +585,7 @@ export const MetricPeekModal: FC<Props> = ({ opened, onClose, metrics }) => {
                                     </Button>
                                 </Group>
 
-                                <MetricPeekComparison
+                                <MetricExploreComparison
                                     baseMetricLabel={metricQuery.data?.label}
                                     query={query}
                                     onQueryChange={setQuery}