+++ packages/frontend/src/features/metricsCatalog/components/MetricsTable.tsx
@@ -34,6 +34,7 @@ import {
     useMemo,
     useRef,
     useState,
+    type FC,
     type UIEvent,
 } from 'react';
 import MantineIcon from '../../../components/common/MantineIcon';
@@ -48,15 +49,21 @@ import {
 import { useMetricsTree } from '../hooks/useMetricsTree';
 import {
     setCategoryFilters,
-    toggleMetricPeekModal,
+    setSearch,
+    setTableSorting,
+    toggleMetricExploreModal,
 } from '../store/metricsCatalogSlice';
 import { MetricCatalogView } from '../types';
-import { MetricPeekModal } from './MetricPeekModal';
+import { MetricExploreModal } from './MetricExploreModal';
 import { MetricsCatalogColumns } from './MetricsCatalogColumns';
 import { MetricsTableTopToolbar } from './MetricsTableTopToolbar';
 import MetricTree from './MetricTree';
 
-export const MetricsTable = () => {
+type MetricsTableProps = {
+    metricCatalogView: MetricCatalogView;
+};
+
+export const MetricsTable: FC<MetricsTableProps> = ({ metricCatalogView }) => {
     const { track } = useTracking();
     const dispatch = useAppDispatch();
     const theme = useMantineTheme();
@@ -73,32 +80,25 @@ export const MetricsTable = () => {
     const { canManageTags, canManageMetricsTree } = useAppSelector(
         (state) => state.metricsCatalog.abilities,
     );
-    const isMetricPeekModalOpen = useAppSelector(
-        (state) => state.metricsCatalog.modals.metricPeekModal.isOpen,
+    const isMetricExploreModalOpen = useAppSelector(
+        (state) => state.metricsCatalog.modals.metricExploreModal.isOpen,
     );
-    const metricCatalogView = useAppSelector(
-        (state) => state.metricsCatalog.view,
+    const search = useAppSelector((state) => state.metricsCatalog.search);
+    const stateTableSorting = useAppSelector(
+        (state) => state.metricsCatalog.tableSorting,
     );
+    const deferredSearch = useDeferredValue(search);
     const prevView = useRef(metricCatalogView);
-
     const tableContainerRef = useRef<HTMLDivElement>(null);
     const rowVirtualizerInstanceRef =
         useRef<MRT_Virtualizer<HTMLDivElement, HTMLTableRowElement>>(null);
-    const [search, setSearch] = useState<string | undefined>(undefined);
-    const deferredSearch = useDeferredValue(search);
 
-    // Enable sorting by highest popularity(how many charts use the metric) by default
-    const initialSorting = [
-        {
-            id: 'chartUsage',
-            desc: true,
-        },
-    ];
+    // We need internal state to handle non serializable state for the updater function
+    const [internalSorting, setInternalSorting] =
+        useState<MRT_SortingState>(stateTableSorting);
 
-    const [sorting, setSorting] = useState<MRT_SortingState>(initialSorting);
-
-    const onCloseMetricPeekModal = () => {
-        dispatch(toggleMetricPeekModal(undefined));
+    const onCloseMetricExploreModal = () => {
+        dispatch(toggleMetricExploreModal(undefined));
     };
 
     const {
@@ -114,11 +114,15 @@ export const MetricsTable = () => {
         search: deferredSearch,
         categories: categoryFilters,
         // TODO: Handle multiple sorting - this needs to be enabled and handled later in the backend
-        ...(sorting.length > 0 && {
-            sortBy: sorting[0].id as keyof CatalogItem,
-            sortDirection: sorting[0].desc ? 'desc' : 'asc',
+        ...(stateTableSorting.length > 0 && {
+            sortBy: stateTableSorting[0].id as keyof CatalogItem,
+            sortDirection: stateTableSorting[0].desc ? 'desc' : 'asc',
         }),
     });
+
+    useEffect(() => {
+        dispatch(setTableSorting(internalSorting));
+    }, [dispatch, internalSorting]);
 
     useEffect(() => {
         if (
@@ -300,13 +304,13 @@ export const MetricsTable = () => {
         enableHiding: false,
         enableGlobalFilterModes: false,
         onGlobalFilterChange: (s: string) => {
-            setSearch(s);
+            dispatch(setSearch(s));
         },
         manualFiltering: true,
         enableFilterMatchHighlighting: true,
         enableSorting: true,
         manualSorting: true,
-        onSortingChange: setSorting,
+        onSortingChange: setInternalSorting,
         enableTopToolbar: true,
         positionGlobalFilter: 'left',
         mantinePaperProps,
@@ -463,7 +467,7 @@ export const MetricsTable = () => {
             <Box>
                 <MetricsTableTopToolbar
                     search={search}
-                    setSearch={setSearch}
+                    setSearch={(s) => dispatch(setSearch(s))}
                     totalResults={totalResults}
                     selectedCategories={categoryFilters}
                     setSelectedCategories={handleSetCategoryFilters}
@@ -472,6 +476,7 @@ export const MetricsTable = () => {
                     showCategoriesFilter={canManageTags || dataHasCategories}
                     isValidMetricsTree={isValidMetricsTree}
                     segmentedControlTooltipLabel={segmentedControlTooltipLabel}
+                    metricCatalogView={metricCatalogView}
                 />
                 <Divider color="gray.2" />
             </Box>
@@ -540,7 +545,7 @@ export const MetricsTable = () => {
             ),
         },
         state: {
-            sorting,
+            sorting: stateTableSorting,
             showProgressBars: false,
             showLoadingOverlay, // show loading overlay when fetching (like search, category filtering), but hide when editing rows.
             showSkeletons: isLoading, // loading for the first time with no data
@@ -595,10 +600,10 @@ export const MetricsTable = () => {
             return (
                 <>
                     <MantineReactTable table={table} />
-                    {isMetricPeekModalOpen && (
-                        <MetricPeekModal
-                            opened={isMetricPeekModalOpen}
-                            onClose={onCloseMetricPeekModal}
+                    {isMetricExploreModalOpen && (
+                        <MetricExploreModal
+                            opened={isMetricExploreModalOpen}
+                            onClose={onCloseMetricExploreModal}
                             metrics={flatData}
                         />
                     )}
@@ -610,7 +615,7 @@ export const MetricsTable = () => {
                     <Box>
                         <MetricsTableTopToolbar
                             search={search}
-                            setSearch={setSearch}
+                            setSearch={(s) => dispatch(setSearch(s))}
                             totalResults={totalResults}
                             selectedCategories={categoryFilters}
                             setSelectedCategories={handleSetCategoryFilters}
@@ -623,6 +628,7 @@ export const MetricsTable = () => {
                             segmentedControlTooltipLabel={
                                 segmentedControlTooltipLabel
                             }
+                            metricCatalogView={metricCatalogView}
                         />
                         <Divider color="gray.2" />
                     </Box>